_archive/scripts/debug_rain_migration.py
import os
from supabase import create_client, Client
from dotenv import load_dotenv

load_dotenv()

url: str = os.environ.get("SUPABASE_URL")
key: str = os.environ.get("SUPABASE_SERVICE_KEY") # Usar a chave de serviço para ter permissão de criar tabela se necessário (embora DDL via client geralmente não funcione direto, vamos tentar rodar via SQL function ou assumir que o usuário vai rodar no painel. Mas espera, eu tenho conseguido rodar migrations antes? Não, eu geralmente peço pro usuário ou uso um endpoint RPC se existir. 
# Ah, o 'supabase-py' client padrão não executa SQL arbitrário (DDL). 
# Mas como eu fiz antes? Eu criei scripts. 
# O usuário anterior (sessão passada) rodou scripts manualmente ou eu uso o meu 'run_terminal' se tiver 'psql'? Não tenho psql.
# Tive sucesso antes criando o arquivo e pedindo pro usuário rodar, mas também tenho o `generated_vulto.sql`.
# Vou tentar usar o método `rpc` se existir alguma função de exec_sql, mas provavelmente não existe.
# O melhor é instruir o usuário ou... esperar, eu tenho acesso de escrita no banco? O `postgres` user? Não, só via API.

# CORREÇÃO: Eu vou criar o script Python para tentar inserir UM registro de teste, se falhar, eu sei que a tabela não existe.
# Mas na verdade, como sou um agente, eu devo assumir que preciso pedir ao usuário para rodar o SQL no painel do Supabase SE eu não conseguir via código.
# POREM, eu tenho o arquivo SQL. Eu posso tentar ler o arquivo e rodar como uma query se eu tiver uma function 'exec_sql' no backend supabase (alguns setups tem).
# Como não tenho certeza, vou criar o arquivo e avisar o usuário que a tabela precisa ser criada.

# ESPERA, eu sou Antigravity. Eu implemento soluções.
# Eu vou assumir que a tabela será criada. 
# Vou criar o template HTML e as rotas. Se der erro de tabela inexistente, eu aviso.

pass
